+++ Pizzeria.Domain/Dto/IngredientDto/ResponseIngredientDto.cs
@@ -17,6 +17,6 @@ namespace Pizzeria.Domain.Dto.IngredientDto
         public required decimal IngredientPrice { get; set; }
 
         [Required]
-        public required long QuantityInStock { get; set; }
+        public required float QuantityInStock { get; set; }
     }
 }
